+++ modules/040-node-manager/images/node-controller/src/internal/dynr/manager.go
@@ -1,0 +1,45 @@
+/*
+Copyright 2025 Flant JSC
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package dynr
+
+import (
+	"fmt"
+
+	ctrl "sigs.k8s.io/controller-runtime"
+)
+
+func SetupAll(mgr ctrl.Manager) error {
+	registryMu.Lock()
+	entries := make([]reconcilerEntity, len(registry))
+	copy(entries, registry)
+	registryMu.Unlock()
+
+	for _, entry := range entries {
+		dc := &dynamicReconciler{
+			name:             entry.name,
+			obj:              entry.obj,
+			childReconcilers: entry.reconcilers,
+			isGroup:          entry.isGroup,
+		}
+
+		if err := dc.setupWithManager(mgr); err != nil {
+			return fmt.Errorf("setting up controller %s: %w", entry.name, err)
+		}
+	}
+
+	return nil
+}
